Responded to Doug's comments
